Home_Work_Project_Доработал задачу. Вынес перевод в рубли в отдельный класс
