+++ src/Home_Work_Project/src/Projects/ConverterTest.java
@@ -2,19 +2,20 @@ package Projects;
 
 import org.junit.jupiter.api.Assertions;
 
-class RubleTest {
+class ConverterTest {
 
     @org.junit.jupiter.api.Test
     void digits2text() {
         String scenario = "Тест преобразования числа";
         try {
-            String actual1 = Ruble.digits2text(3.0);
+
+            String actual1 = Converter.digits2text(3.0);
             Assertions.assertEquals("Три рубля 00 копеек", actual1);
 
-            String actual2 = Ruble.digits2text(33.0);
+            String actual2 = Converter.digits2text(33.0);
             Assertions.assertEquals("Тридцать три рубля 00 копеек", actual2);
 
-            String actual3 = Ruble.digits2text(333.0);
+            String actual3 = Converter.digits2text(333.0);
             Assertions.assertEquals("Триста тридцать три рубля 00 копеек", actual3);
 
             System.out.printf("\"%s\" passed %n", scenario);
