+++ src/Home_Work_Project/src/Projects/Currency.java
@@ -1,0 +1,8 @@
+package Projects;
+
+public interface Currency {
+
+    String addRub(Double d, String input);
+
+
+}
