+++ src/Home_Work_Project/src/Projects/Main.java
@@ -8,6 +8,16 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         int number = Integer.parseInt(reader.readLine());
-        System.out.println(Ruble.digits2text((double) number));
+        //System.out.println(new Ruble().addRub((double) number, Converter.digits2text((double) number)));
+
+        convert(number, new Ruble());
+
+
+
+    }
+
+    static String convert(int digit, Currency currency){
+        System.out.println(currency.addRub((double) digit, Converter.digits2text((double) digit)));
+        return null;
     }
 }
